clnrest: refactor `POST /v1/<rpc_method>` requests handling
- remove function `verify_rune`, we directly call
  `plugin.rpc.call("checkrune", ...)`
- remove function `call_rpc_method`, we directly call
  `plugin.rpc.call(...)`.

A cool 'side effect' of removing those functions is that we no
longer have to check twice for the parameters of the rpc method
in the request.

We also take advantage of the fact that errors raised when
`plugin.rpc.call(...)` failed are of type `RpcError` and we can access
the error with the field `error`, no need to `json5.loads(...)`
on the error or to parse it looking for the field "error" in it.